feat: add lightweight text layout helper
